requirements.txt
uvicorn~=0.20.0
fastapi~=0.89.1
websocket-client~=0.59.0
cryptography~=39.0.0
gunicorn
websockets
loguru
